Version 2022.07.08-02
remove paths.py : no more used,
add rootValues.py and modify createRootFiles.sh
rootValues.py contains the values to be used with the createRootFiles.sh script. This help for the creation of a ROOT file with NB_EVTS events from number Nbegin to number Nend.
NOTE : Nbegin..Nend are the numbers of each ROOT file and NOT the number of events inside the NB_EVTS events.