Version 2022.08.11-02
checkMapDiff tool : OK. Generate map of the diffs between 2 ROOT files for each hsitogram and can show a pbm into ROOT file.
add a new tool (statpValues.py) in order to generate some graphs reporting the number of histograms as a function of the pValue. This can show some information of the validity of the release.