Version 2023.05.30-01
Correction of pValue display. A regression was found with decisionBoxX (X=1,2,3) with a false value of diffKS. We need to load a value from a text file to be OK.
Next modifications to avoid those errors have to be regarded.
The call to diffMaxKS3 get an error and need to be treated later.