Version 2023.06.06-01
add a new file (to be included into resumeAE.py ?) in order to compare the number of recomputed with AE histos for each release and for each scheme (with a ndifferent number of layers).
This comparison is simple with nb sup to 5% (red) and nb inf to 5% (green). Before this, the losses values have been normalized to 1 for the max.